# swagger.yaml
openapi: 3.0.0
info:
  title: Node Tracker API
  version: 1.0.0
  description: REST API for a private BitTorrent tracker with user management, torrents, and IP bans.
  contact:
    name: API Support
    email: support@nodetracker.com
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Torrent:
      type: object
      properties:
        id:
          type: integer
          description: Unique torrent ID
        name:
          type: string
          description: Torrent name
        description:
          type: string
          description: Torrent description
        infoHash:
          type: string
          description: Unique torrent hash
        size:
          type: integer
          description: Size in bytes
        completed:
          type: integer
          description: Number of completed downloads
        downloads:
          type: integer
          description: Total number of downloads
        anonymous:
          type: boolean
          description: Whether the torrent is anonymous
        freeleech:
          type: boolean
          description: Whether the torrent is freeleech
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
security:
  - bearerAuth: []
...
paths:
...
  /api/torrents:
    get:
      summary: List all torrents with pagination and filters
      tags: [Torrents]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results limit per page
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
        - in: query
          name: search
          schema:
            type: string
          description: Search by name
      responses:
        200:
          description: Torrent list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  torrents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Torrent'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create a new torrent
      tags: [Torrents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [infoHash, name]
              properties:
                infoHash:
                  type: string
                  minLength: 40
                  maxLength: 40
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                category:
                  type: string
                tags:
                  type: string
                size:
                  type: integer
                  minimum: 0
                anonymous:
                  type: boolean
                freeleech:
                  type: boolean
      responses:
        201:
          description: Torrent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torrent'
  /api/torrents/{id}:
    get:
      summary: Get specific torrent information
      tags: [Torrents]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Torrent ID
      responses:
        200:
          description: Torrent information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torrent'

    put:
      summary: Update a torrent
      tags: [Torrents]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Torrent ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                category:
                  type: string
                tags:
                  type: string
                anonymous:
                  type: boolean
                freeleech:
                  type: boolean
      responses:
        200:
          description: Torrent updated successfully

    delete:
      summary: Delete a torrent
      tags: [Torrents]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Torrent ID
      responses:
        200:
          description: Torrent deleted successfully
        404:
          description: Torrent not found
...